+++ frontend/realestate/src/browseAll/NavBottom.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "./style.css";
+import "../style.css";
 
 const setting = {
   textAlign: "center",
