+++ frontend/realestate/src/browseOne/BrowseItemBottomLine.js
